newlinter: Adding gofuncor linter
