--- pkg/golinters/gofuncor/testdata/struct_not_declared_in_file.go
@@ -1,13 +1,0 @@
-package simple
-
-import (
-	"time"
-)
-
-func NewOtherWayMyStruct() MyStruct {
-	return MyStruct{Name: "John"}
-}
-
-func NewTimeStruct() time.Time {
-	return time.Now()
-}
